# Conflicts:
#	data/systems/00_sol.lua
#	src/terrain/Makefile.am
#	src/terrain/Terrain.cpp
#	src/terrain/Terrain.h
#	src/terrain/TerrainColorAsteroid.cpp
#	src/terrain/TerrainColorBandedRock.cpp
#	src/terrain/TerrainColorDeadWithWater.cpp
#	src/terrain/TerrainColorDesert.cpp
#	src/terrain/TerrainColorEarthLike.cpp
#	src/terrain/TerrainColorEarthLikeHeightmapped.cpp
#	src/terrain/TerrainColorGGJupiter.cpp
#	src/terrain/TerrainColorGGNeptune.cpp
#	src/terrain/TerrainColorGGNeptune2.cpp
#	src/terrain/TerrainColorGGSaturn.cpp
#	src/terrain/TerrainColorGGSaturn2.cpp
#	src/terrain/TerrainColorGGUranus.cpp
#	src/terrain/TerrainColorIce.cpp
#	src/terrain/TerrainColorMethane.cpp
#	src/terrain/TerrainColorRock.cpp
#	src/terrain/TerrainColorRock2.cpp
#	src/terrain/TerrainColorStarBrownDwarf.cpp
#	src/terrain/TerrainColorStarG.cpp
#	src/terrain/TerrainColorStarK.cpp
#	src/terrain/TerrainColorStarM.cpp
#	src/terrain/TerrainColorStarWhiteDwarf.cpp
#	src/terrain/TerrainColorTFGood.cpp
#	src/terrain/TerrainColorTFPoor.cpp
#	src/terrain/TerrainColorVolcanic.cpp
#	src/terrain/TerrainFeature.cpp
#	src/terrain/TerrainFeature.h
#	src/terrain/TerrainHeightAsteroid.cpp
#	src/terrain/TerrainHeightAsteroid2.cpp
#	src/terrain/TerrainHeightAsteroid3.cpp
#	src/terrain/TerrainHeightAsteroid4.cpp
#	src/terrain/TerrainHeightBarrenRock.cpp
#	src/terrain/TerrainHeightBarrenRock2.cpp
#	src/terrain/TerrainHeightBarrenRock3.cpp
#	src/terrain/TerrainHeightEllipsoid.cpp
#	src/terrain/TerrainHeightFlat.cpp
#	src/terrain/TerrainHeightHillsCraters.cpp
#	src/terrain/TerrainHeightHillsCraters2.cpp
#	src/terrain/TerrainHeightHillsDunes.cpp
#	src/terrain/TerrainHeightHillsNormal.cpp
#	src/terrain/TerrainHeightHillsRidged.cpp
#	src/terrain/TerrainHeightHillsRivers.cpp
#	src/terrain/TerrainHeightMapped.cpp
#	src/terrain/TerrainHeightMapped2.cpp
#	src/terrain/TerrainHeightMountainsCraters.cpp
#	src/terrain/TerrainHeightMountainsCraters2.cpp
#	src/terrain/TerrainHeightMountainsNormal.cpp
#	src/terrain/TerrainHeightMountainsRidged.cpp
#	src/terrain/TerrainHeightMountainsRivers.cpp
#	src/terrain/TerrainHeightMountainsRiversVolcano.cpp
#	src/terrain/TerrainHeightMountainsVolcano.cpp
#	src/terrain/TerrainHeightRuggedDesert.cpp
#	src/terrain/TerrainHeightRuggedLava.cpp
#	src/terrain/TerrainHeightWaterSolid.cpp
#	src/terrain/TerrainHeightWaterSolidCanyons.cpp
#	win32/vs2015/terrain/terrain.vcxproj
#	win32/vs2015/terrain/terrain.vcxproj.filters

+++ src/LuaTable.h
@@ -1,4 +1,4 @@
-// Copyright © 2008-2017 Pioneer Developers. See AUTHORS.txt for details
+// Copyright © 2008-2018 Pioneer Developers. See AUTHORS.txt for details
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #ifndef _LUATABLE_H
@@ -7,7 +7,8 @@
 #include <cassert>
 #include <iterator>
 
-#include "lua/lua.hpp"
+#include <lua.hpp>
+
 #include "LuaRef.h"
 #include "LuaPushPull.h"
 #include "LuaUtils.h"
@@ -141,7 +142,7 @@ public:
 
 	lua_State * GetLua() const { return m_lua; }
 	int GetIndex() const { return m_index; }
-	int Size() const {return lua_rawlen(m_lua, m_index);}
+	size_t Size() const {return lua_rawlen(m_lua, m_index);}
 
 	/* VecIter, as in VectorIterator (only shorter to type :-)
 	 *
